feat: adds a JDialog confirmation before deleting a player from the DB

+++ src/org/krauss/gui/admin/PanelAdmin.java
@@ -13,6 +13,7 @@ import javax.swing.DefaultListModel;
 import javax.swing.JButton;
 import javax.swing.JLabel;
 import javax.swing.JList;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JPasswordField;
 import javax.swing.JScrollPane;
@@ -26,30 +27,30 @@ import net.miginfocom.swing.MigLayout;
  * 
  * @author jrkrauss <br>
  *         <br>
- *         This class defines the panel of the admin interface. All the components are
- *         setup in this class.
+ *         This class defines the panel of the admin interface. All the
+ *         components are setup in this class.
  *
  */
 @SuppressWarnings("serial")
 public class PanelAdmin extends JPanel {
-	
-	//Database
+
+	// Database
 	private static DatabaseHandler dbHandler;
 	private ArrayList<Player> rows;
 	private Object[][] data;
-	
-	//Option panel components
+
+	// Option panel components
 	private JPanel jp_optionsPanel;
 
-	//Database Table components
+	// Database Table components
 	private JScrollPane jsp_scrollPane;
 	private JTable jt_rankTable;
-	private String collumns[] = {"Player", "Score", "Last Login"};
-	
-	//Rank List components
+	private String collumns[] = { "Player", "Score", "Last Login" };
+
+	// Rank List components
 	private JList<Object> jli_rank;
-	
-	//Options components
+
+	// Options components
 	private JLabel jl_player;
 	private JLabel jl_playerData;
 	private JButton jb_delete;
@@ -59,23 +60,22 @@ public class PanelAdmin extends JPanel {
 	private JButton jb_edit;
 
 	public PanelAdmin() {
-		
 		this.setLayout(new MigLayout("", "10[]5[]5[]10", "10[grow]10"));
-		this.setBorder(BorderFactory.createLineBorder(Color.WHITE, 5, true));			
-		
+		this.setBorder(BorderFactory.createLineBorder(Color.WHITE, 5, true));
+
 		createJPanelOptins();
 		createJPanelDatabase();
 		createJPanelRank();
 
 	}
-	
+
 	private void createJPanelOptins() {
-		
+
 		jp_optionsPanel = new JPanel();
 		jp_optionsPanel.setLayout(new MigLayout("", "10[]7[]10", "15[]7[]7[]10[]10"));
 		jp_optionsPanel.setPreferredSize(new Dimension(200, 490));
 		jp_optionsPanel.setBorder(BorderFactory.createTitledBorder("Options"));
-		
+
 		jl_player = new JLabel("Player:");
 		jl_player.setPreferredSize(new Dimension(100, 29));
 		jl_playerData = new JLabel();
@@ -92,27 +92,28 @@ public class PanelAdmin extends JPanel {
 		jb_edit.setPreferredSize(new Dimension(190, 29));
 		jb_edit.addActionListener(new EditButtonListener());
 		jb_edit.setEnabled(false);
-		
+
 		jl_player.setEnabled(false);
 		jl_playerData.setEnabled(false);
-		
+
 		jp_optionsPanel.add(jl_player, "cell 0 0, left");
 		jp_optionsPanel.add(jl_playerData, "cell 1 0, left");
-		/*jp_optionsPanel.add(jb_resetPasswd, "cell 0 1, span 2, growx, left");
-		jp_optionsPanel.add(jb_delete, "cell 0 2, left");
-		jp_optionsPanel.add(jb_cancel, "cell 1 2, left");
-		jp_optionsPanel.add(jb_edit, "cell 0 3, span 2, growx, left");*/
+		/*
+		 * jp_optionsPanel.add(jb_resetPasswd, "cell 0 1, span 2, growx, left");
+		 * jp_optionsPanel.add(jb_delete, "cell 0 2, left");
+		 * jp_optionsPanel.add(jb_cancel, "cell 1 2, left");
+		 * jp_optionsPanel.add(jb_edit, "cell 0 3, span 2, growx, left");
+		 */
 		jp_optionsPanel.add(jb_edit, "cell 0 3, span 2, left");
-		
-		
+
 		this.add(jp_optionsPanel, "cell 0 0, center");
 	}
-	
+
 	private void createJPanelDatabase() {
-		
+
 		JPanel jp_rankPanel = new JPanel();
 		jp_rankPanel.setLayout(new MigLayout());
-		
+
 		jt_rankTable = new JTable(fetchDataBaseData(), collumns);
 		jt_rankTable.setModel(new DefaultTableModel(fetchDataBaseData(), collumns));
 		jt_rankTable.setFillsViewportHeight(true);
@@ -120,70 +121,66 @@ public class PanelAdmin extends JPanel {
 		jt_rankTable.setToolTipText("Select a player to enable the options");
 		jsp_scrollPane = new JScrollPane(jt_rankTable);
 		jsp_scrollPane.setPreferredSize(new Dimension(360, 490));
-		
+
 		jt_rankTable.getColumnModel().getColumn(0).setPreferredWidth(120);
 		jt_rankTable.getColumnModel().getColumn(1).setPreferredWidth(70);
 		jt_rankTable.getColumnModel().getColumn(2).setPreferredWidth(160);
-		
+
 		jt_rankTable.setIntercellSpacing(new Dimension(3, 3));
-		
+
 		jt_rankTable.addMouseListener(new TableListener());
-		
-		
+
 		jp_rankPanel.add(jt_rankTable.getTableHeader(), "cell 0 0, center");
 		jp_rankPanel.add(jsp_scrollPane, "cell 0 1, center");
 
 		this.add(jp_rankPanel, "cell 1 0, center");
-		
+
 	}
-	
-	
+
 	private void createJPanelRank() {
-		
+
 		JPanel jp_rankPanel = new JPanel();
 		jp_rankPanel.setLayout(new MigLayout("", "5[]5", "15[]5"));
 		jp_rankPanel.setPreferredSize(new Dimension(220, 490));
 		jp_rankPanel.setBorder(BorderFactory.createTitledBorder("Ranking"));
-		
+
 		@SuppressWarnings("unchecked")
 		ArrayList<Player> rankPlayers = (ArrayList<Player>) rows.clone();
 		rankPlayers.sort(Comparator.comparing(Player::getScore).reversed());
-		
+
 		DefaultListModel<Object> m = new DefaultListModel<Object>();
 		m.addAll(rankPlayers);
 		jli_rank = new JList<Object>(m);
-				
+
 		jli_rank.setEnabled(false);
 		jli_rank.setBackground(jp_rankPanel.getBackground());
 		JScrollPane rank_listScroller = new JScrollPane(jli_rank);
 		rank_listScroller.setPreferredSize(new Dimension(200, 470));
-		
+
 		jp_rankPanel.add(rank_listScroller, "cell 0 0, center");
-				
+
 		this.add(jp_rankPanel, "cell 2 0, center");
 	}
-	
-	
+
 	private Object[][] fetchDataBaseData() {
-		
+
 		dbHandler = new DatabaseHandler();
-		rows = dbHandler.getDBData(); 
+		rows = dbHandler.getDBData();
 		data = new Object[rows.size()][];
-		
-		for (int i = 0; i < rows.size(); i++ ) {
-			
+
+		for (int i = 0; i < rows.size(); i++) {
+
 			data[i] = rows.get(i).playerToString().split(",");
-		} 
-		
+		}
+
 		return data;
-		
+
 	}
-	
-	
+
 	private void updateJTableData(int row) {
-		
+
 		if (jt_rankTable.getModel() instanceof DefaultTableModel) {
-			((DefaultTableModel)jt_rankTable.getModel()).removeRow(row);
+			((DefaultTableModel) jt_rankTable.getModel()).removeRow(row);
 			dbHandler = new DatabaseHandler();
 			rows = dbHandler.getDBData();
 			jt_rankTable.updateUI();
@@ -195,19 +192,17 @@ public class PanelAdmin extends JPanel {
 			jp_optionsPanel.remove(jb_cancel);
 			jp_optionsPanel.updateUI();
 		}
-		
-		
+
 	}
-	
-	
+
 	private void updateJListRank(String playername) {
-		
+
 		Object[] list = ((DefaultListModel<Object>) jli_rank.getModel()).toArray();
-			
+
 		for (int i = 0; i < list.length; i++) {
 			if (list[i] instanceof Player) {
-				
-				if (((Player)list[i]).getPlayerName().equalsIgnoreCase(playername)) {
+
+				if (((Player) list[i]).getPlayerName().equalsIgnoreCase(playername)) {
 					((DefaultListModel<Object>) jli_rank.getModel()).remove(i);
 					break;
 				}
@@ -215,18 +210,17 @@ public class PanelAdmin extends JPanel {
 		}
 		jli_rank.setEnabled(false);
 		jli_rank.updateUI();
-		
+
 	}
-	
-	
+
 	/**
 	 * 
 	 * @author jrkrauss
-	 *	
-	 *	Listener private classes
+	 * 
+	 *         Listener private classes
 	 *
-	 */	
-	private class TableListener implements MouseListener{
+	 */
+	private class TableListener implements MouseListener {
 
 		@Override
 		public void mouseClicked(MouseEvent e) {
@@ -235,34 +229,38 @@ public class PanelAdmin extends JPanel {
 				JTable jt = (JTable) e.getSource();
 				if (jt.getSelectedRow() >= 0) {
 					jl_playerData.setText(rows.get(jt.getSelectedRow()).getPlayerName());
-					//Enables the Edit button only if it's not the admin account.
-					if(jl_playerData.getText().equalsIgnoreCase("admin")) {						
+					// Enables the Edit button only if it's not the admin account.
+					if (jl_playerData.getText().equalsIgnoreCase("admin")) {
 						jb_edit.setEnabled(false);
 					} else {
 						jb_edit.setEnabled(true);
 					}
-					
+
 				}
-				
-			} 
-			
+
+			}
+
 		}
 
 		@Override
-		public void mousePressed(MouseEvent e) {}
+		public void mousePressed(MouseEvent e) {
+		}
 
 		@Override
-		public void mouseReleased(MouseEvent e) {}
+		public void mouseReleased(MouseEvent e) {
+		}
 
 		@Override
-		public void mouseEntered(MouseEvent e) {}
+		public void mouseEntered(MouseEvent e) {
+		}
 
 		@Override
-		public void mouseExited(MouseEvent e) {}
-		
+		public void mouseExited(MouseEvent e) {
+		}
+
 	}
 
-	private class EditButtonListener implements ActionListener{
+	private class EditButtonListener implements ActionListener {
 
 		@Override
 		public void actionPerformed(ActionEvent e) {
@@ -273,10 +271,10 @@ public class PanelAdmin extends JPanel {
 			jp_optionsPanel.add(jb_cancel, "cell 1 2, left");
 			jp_optionsPanel.updateUI();
 		}
-		
+
 	}
-	
-	private class CancelButtonListener implements ActionListener{
+
+	private class CancelButtonListener implements ActionListener {
 
 		@Override
 		public void actionPerformed(ActionEvent e) {
@@ -284,49 +282,59 @@ public class PanelAdmin extends JPanel {
 			jl_playerData.setEnabled(false);
 			jb_resetPasswd.setText("Reset Password");
 			jb_resetPasswd.setForeground(Color.BLACK);
-			if(jpf_password.isVisible()) {
+			if (jpf_password.isVisible()) {
 				jp_optionsPanel.remove(jpf_password);
-			}			
+			}
 			jp_optionsPanel.remove(jb_resetPasswd);
 			jp_optionsPanel.remove(jb_delete);
 			jp_optionsPanel.remove(jb_cancel);
 			jp_optionsPanel.updateUI();
 		}
-		
+
 	}
-	
-	private class DeleteButtonListener implements ActionListener{
+
+	private class DeleteButtonListener implements ActionListener {
 
 		@Override
 		public void actionPerformed(ActionEvent e) {
-			//Open the DB connection
-			dbHandler = new DatabaseHandler();
-			//Delete the player
-			dbHandler.deletePlayer(jl_playerData.getText());
+
+			Object[] options = { "Yeah", "Noou" };
+			int decision = JOptionPane.showOptionDialog(null, "Mate, are you sure you wanna get rid of this bloke?",
+					"Confirmation", JOptionPane.YES_NO_OPTION, JOptionPane.WARNING_MESSAGE, null, options, options[1]);
 			
-			//Find his index within the ArrayList rows
-			int index = 0;
-			
-			for(int i = 0; i < rows.size(); i++) {
-				 if (rows.get(i).getPlayerName().equalsIgnoreCase(jl_playerData.getText())) {
-					index = i;
-					break;
-				 }
+			//Delete the user only if he picks the option 'Yeah'
+			if (decision == 0) {
+
+				// Open the DB connection
+				dbHandler = new DatabaseHandler();
+				// Delete the player
+				dbHandler.deletePlayer(jl_playerData.getText());
+
+				// Find his index within the ArrayList rows
+				int index = 0;
+
+				for (int i = 0; i < rows.size(); i++) {
+					if (rows.get(i).getPlayerName().equalsIgnoreCase(jl_playerData.getText())) {
+						index = i;
+						break;
+					}
+				}
+				updateJListRank(jl_playerData.getText());
+				updateJTableData(index);
+
 			}
-			updateJListRank(jl_playerData.getText());
-			updateJTableData(index);
-			
+
 		}
-		
+
 	}
-	
-	private class ResetButtonListener implements ActionListener{
+
+	private class ResetButtonListener implements ActionListener {
 
 		@Override
 		public void actionPerformed(ActionEvent e) {
-			//Open the DB connection
-			if (((JButton)e.getSource()).getText().equalsIgnoreCase("Reset Password")) {
-				
+			// Open the DB connection
+			if (((JButton) e.getSource()).getText().equalsIgnoreCase("Reset Password")) {
+
 				dbHandler = new DatabaseHandler();
 				jpf_password.setText("");
 				jp_optionsPanel.add(jpf_password, "cell 0 1, span 2, growx, left");
@@ -337,9 +345,9 @@ public class PanelAdmin extends JPanel {
 				jb_resetPasswd.setText("Apply");
 				jb_resetPasswd.setForeground(new Color(0x03853E));
 				jp_optionsPanel.updateUI();
-				
-			} else if (((JButton)e.getSource()).getText().equalsIgnoreCase("Apply")) {
-				
+
+			} else if (((JButton) e.getSource()).getText().equalsIgnoreCase("Apply")) {
+
 				dbHandler.updatePlayerPassword(jl_playerData.getText(), jpf_password.getPassword());
 				jb_resetPasswd.setText("Reset Password");
 				jb_resetPasswd.setForeground(Color.BLACK);
@@ -349,13 +357,9 @@ public class PanelAdmin extends JPanel {
 				jp_optionsPanel.remove(jb_cancel);
 				jp_optionsPanel.updateUI();
 			}
-			
-			
-			
-						
+
 		}
-		
+
 	}
-	
-	
+
 }
